Deprecate writing model to pickle - move functions to general model
