+++ pyproject.toml
@@ -55,7 +55,7 @@ extend-select = [
 ignore = ["D100", "D101", "D415", "ANN101", "ANN102", "ANN002", "ANN003"]
 
 [tool.ruff.lint.extend-per-file-ignores]
-"tests/unit/*.py" = ["ANN201", "D102", "D103"]
+"tests/unit/*.py" = ["ANN201", "D102", "D103", "D106"]
 
 [tool.ruff.lint.flake8-comprehensions]
 allow-dict-calls-with-keyword-arguments = true
